// nextjs/Interface.ts
export interface EnvironmentVariable {
	id: string;
	key: string;
	value: string;
}